refactored code to improve reusability and maintainability
